+++ Sources/Interface/interfaceBook.py
@@ -1,0 +1,44 @@
+from settings import * 
+
+class BookInterface(object):
+
+    def __init__(self, gestionnaire: any) ->None:
+        """Méthode initialisation de l'interface de book.
+        Input : gestionnaire = self méthode d'appel"""
+        
+        # Initialisation
+        self.gestionnaire = gestionnaire
+        
+        # Création éléments
+        self.displaySurface = pygame.display.get_surface() # surface générale
+        self.interfaceSurface = pygame.Surface((WINDOW_WIDTH/2, WINDOW_HEIGHT/2),  pygame.SRCALPHA)
+        self.interfaceSurface.fill("#ffffff")
+
+
+    def BuildInterface(self) -> None:
+        """Méthode : Création de tout les éléments composant l'interface. Input / Output : None"""
+
+        # texte titre
+        self.interfaceSurface.fill("#ffffff")
+        text = FONT["FONT36"].render(TEXTE["Elements"]["HotBar"]["Book"]["Title"], True, (0,0,0))
+        self.interfaceSurface.blit(text, (10,10))
+
+
+    def CloseInterface(self) -> None:
+        """Méthode de fermeture de l'interface. Input / Output : None"""
+
+        # changement des boolean de check
+        self.gestionnaire.CloseInterface()
+
+
+    def Update(self, event) -> None:
+        """Méthode d'update de l'interface. Input / Output : None"""
+
+        # construction d'update
+        self.BuildInterface()
+        self.displaySurface.blit(self.interfaceSurface, (320,180)) # pos topleft
+
+        # Fermer l'interface avec ESC
+        keys = pygame.key.get_pressed()
+        if keys[KEYSBIND["echap"]]:  # Fermer avec ESC
+            self.CloseInterface()
