+++ Sources/Interface/interfaceExo.py
@@ -159,7 +159,7 @@ class CreateExo:
         """Méthode de fermeture de l'interface. Input / Output : None"""
 
         # changement des boolean de check
-        self.gestionnaire.INTERFACE_OPEN = False
+
         self.gestionnaire.interface_exo = False
         INFOS["Exo"] = False
