+++ Sources/Interface/interfaceReactor.py
@@ -1,0 +1,51 @@
+from settings import * 
+
+class ReactorInterface(object):
+
+    def __init__(self) ->None:
+        """Méthode initialisation de l'interface de book.
+        Input : gestionnaire = self méthode d'appel"""
+           
+        # Création éléments
+        self.displaySurface = pygame.display.get_surface() # surface générale
+
+        # Surface pour le fond transparent
+        self.backgroundSurface = pygame.Surface((WINDOW_WIDTH, WINDOW_HEIGHT), pygame.SRCALPHA)
+        self.backgroundSurface.fill((50, 50, 50, 200))  # Fond gris avec alpha (200)
+
+        # Surface principale pour les éléments de l'interface
+        self.interfaceSurface = pygame.Surface((WINDOW_WIDTH, WINDOW_HEIGHT), pygame.SRCALPHA)
+        self.interfaceSurface.fill((0, 0, 0, 0))  # Transparent par défaut
+
+        # timer click skip
+        self.last_click_time = 0
+        self.click_delay = 500  
+
+
+    def BuildInterface(self) -> None:
+        """Méthode : Création de tout les éléments composant l'interface. Input / Output : None"""
+
+        # texte titre
+        self.interfaceSurface.fill("#ffffff")
+        text = FONT["FONT36"].render(TEXTE["Elements"]["HotBar"]["Book"]["Title"], True, (0,0,0))
+        self.interfaceSurface.blit(text, (10,10))
+
+
+    def CloseInterface(self) -> None:
+        """Méthode de fermeture de l'interface. Input / Output : None"""
+
+        # changement des boolean de check
+        self.gestionnaire.CloseInterface()
+        
+
+    def Update(self, event) -> None:
+        """Méthode d'update de l'interface. Input / Output : None"""
+
+        # construction d'update
+        self.BuildInterface()
+        self.displaySurface.blit(self.interfaceSurface, (320,180)) # pos topleft
+
+        # Fermer l'interface avec ESC
+        keys = pygame.key.get_pressed()
+        if keys[KEYSBIND["echap"]]:  # Fermer avec ESC
+            self.CloseInterface()
